+++ packages/common-substrate/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `isSigned` filter to call handler (#1940)
 
 ## [2.3.0] - 2023-08-10
 ### Changed